[web] Split the Rust library into modules.
